Uploading file that parses TAF into hourly chunks
This script now takes the TAF and parses the complete TAF into the hourly representation of the TAF, this will aid in helping us what is happening and how we can get verification scores from it. 

Also, I uploaded the new data.json output so people can see the change.